+++ Changelog.md
@@ -3,6 +3,10 @@
 1. Use zero copy when publishing messages.
     * [Pull request 229](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/229)
 
+1. Updated service responder callbacks to return a boolean value. The
+   existing functions have been deprecated.
+    * [Pull request 260](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/260)
+
 ## Ignition Transport 3.0.0
 
 1. Added optional message throttling when publishing messages.